Revert "[0373] Optimise course search query "
