+++ spec/services/provider_reporting_service_spec.rb
@@ -93,10 +93,10 @@ describe ProviderReportingService do
         expect(providers_scope).to receive(:distinct).and_return(distinct_providers_scope)
         expect(distinct_providers_scope).to receive(:count).and_return(providers_count)
         expect(distinct_providers_scope).to receive(:count).and_return(providers_count)
-        expect(distinct_providers_scope).to receive(:where).with(id: Course.findable.pluck(:provider_id)).and_return(training_providers_scope)
+        expect(distinct_providers_scope).to receive(:where).with(id: Course.findable.select(:provider_id)).and_return(training_providers_scope)
 
         expect(distinct_providers_scope).to receive(:where)
-          .with(id: Course.findable.with_vacancies.pluck(:provider_id))
+          .with(id: Course.findable.with_vacancies.select(:provider_id))
           .and_return(open_providers_scope)
 
         expect(training_providers_scope).to receive_message_chain(:where, :not).and_return(closed_providers_scope)
